Modified stack to move storage user from shared to app stack.
AWS IAM limits the number of access keys to 2, rather unexpectedly.
Having the user for the S3 bucket in the stack controlling it seems
logical, but the secret key can only be accessed at credentials creation
time. That means we'd need to either export the secret key in the clear,
or move the user creation to the app stack, the latter being judged less
bad.