Enable a public bucket on moj-wp-prod
This is a short-term measure to address the problem of the documents
that were deleted when the old prod bucket was removed. I have chosen to
make it public website as the documents are not sensitive and it
simplifies bucket management.

For this site to work, the documents must be stored in the
`/imported_content` path. The docuemnts have been restored from a copy
that was imported into the current dev bucket.

Again, this is only a short-term solution designed to reduce the number
of support queries arising from the missing documents while we find a
better way to resolve the issue.  This checkin **does not** copy the
documents, it only enables the bucket.  The documents will need to be
copied manually.