Add a parameter to choose which branch to deploy
In most cases, we want to deploy the `master` branch of our application from the repository. However for PPJ we have a second branch which we want to deploy to a separate pipeline. We'll use this to develop a new version of the site in a separate branch, whilst maintaining our ability to use the existing pipeline for patches and fixes to the current live site. Eventually this branch will be merged into master and pushed to through the production pipeline.

In order to allow this workflow, I've added a parameter to the deployment pipeline CloudFormation template to allow us to specify which branch of the repository to deploy. It defaults to master, so is backwards compatible with existing pipeline templates.