Add stack parameters for maintenance mode
I've introduced two parameters to the hosting template, which are to be passed through as environment variables to the docker container. Parameter `MaintenanceMode` becomes env var `MAINTENANCE_MODE`, and `MaintenanceWhitelistIps` becomes `MAINTENANCE_WHITELIST_IPS`.

These are to be used to enable 'maintenance mode' – which, if implemented, can be used to display a maintenance page to end users, meanwhile allowing access to whitelisted IP addresses. Note that it is the responsibility of the container to consume the environment variables and implement the described functionality – this is beyond the scope of this repository. This change simply makes it possible.

Maintenance page functionality has been implemented in ministryofjustice/wp-ppj, so this may be used as a reference.